reorganise SelectiveRetirement files
